auth_demo/requirements.txt
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
python-dotenv==0.19.2
social-auth-app-django==5.0.0
